- make ignoring of function results explicit or handle it. - move assigments out of conditions in "while" or "if" statements - reduce implicit signed/unsigned conversions

+++ nsd/range.c
@@ -132,7 +132,7 @@ NsConnParseRange(Ns_Conn *conn, const char *type,
     }
 
     maxranges = (*nbufsPtr / 2) - 1;
-    ranges = alloca(sizeof(Range) * maxranges);
+    ranges = alloca(sizeof(Range) * (size_t)maxranges);
 
     rangeCount = ParseRangeOffsets(conn, objLength, ranges, maxranges);
     if (rangeCount < 1) {
